Move views/views/*.vue -> views/*.vue
