+++ frontend/simple/views/Invite.vue
@@ -43,13 +43,13 @@
 }
 </style>
 <script>
-import * as Events from '../../../../shared/events'
-import backend from '../../controller/backend/'
-import L from '../utils/translations'
+import * as Events from '../../../shared/events'
+import backend from '../controller/backend/'
+import L from './utils/translations'
 import template from 'string-template'
-import { GroupInvitees } from '../components/CreateGroupSteps'
+import { GroupInvitees } from './components/CreateGroupSteps'
 import { mapState, mapGetters } from 'vuex'
-import sbp from '../../../../shared/sbp'
+import sbp from '../../../shared/sbp'
 
 export default {
   name: 'Invite',
