+++ frontend/simple/views/Mailbox.vue
@@ -195,11 +195,11 @@
 </style>
 <script>
 import _ from 'lodash'
-import backend from '../../controller/backend'
-import * as Events from '../../../../shared/events'
-import {namespace} from '../../controller/backend/hapi'
-import {latestContractState} from '../../model/state'
-import L from '../utils/translations'
+import backend from '../controller/backend'
+import * as Events from '../../../shared/events'
+import {namespace} from '../controller/backend/hapi'
+import {latestContractState} from '../model/state'
+import L from './utils/translations'
 const criteria = [(msg) => new Date(msg.sentDate)]
 export default {
   name: 'Mailbox',
