+++ frontend/simple/views/SignUp.vue
@@ -110,15 +110,15 @@
 }
 </style>
 <script>
-import backend from '../../controller/backend'
+import backend from '../controller/backend'
 import Vue from 'vue'
 import _ from 'lodash'
-import * as Events from '../../../../shared/events'
-import * as contracts from '../../model/contracts/events'
-import {namespace} from '../../controller/backend/hapi'
+import * as Events from '../../../shared/events'
+import * as contracts from '../model/contracts/events'
+import {namespace} from '../controller/backend/hapi'
 import { validationMixin } from 'vuelidate'
 import { required, minLength, email } from 'vuelidate/lib/validators'
-import { nonWhitespace } from '../utils/validators'
+import { nonWhitespace } from './utils/validators'
 // TODO: fix all this
 export default {
   name: 'SignUp',
