+++ frontend/simple/views/UserGroup.vue
