+++ frontend/simple/views/UserProfile.vue
@@ -219,13 +219,13 @@
   </section>
 </template>
 <script>
-import backend from '../../controller/backend'
-import * as Events from '../../../../shared/events'
-import L from '../utils/translations'
+import backend from '../controller/backend'
+import * as Events from '../../../shared/events'
+import L from './utils/translations'
 import _ from 'lodash'
 import { validationMixin } from 'vuelidate'
 import { url, email } from 'vuelidate/lib/validators'
-import { decimals } from '../utils/validators'
+import { decimals } from './utils/validators'
 
 export default {
   name: 'UserProfile',
